upgraded to nextjs14 and live preview 2.0
