+++ .env.local.sample
@@ -7,9 +7,9 @@ CONTENTSTACK_DELIVERY_TOKEN=YOUR_DELIVERY_TOKEN
 CONTENTSTACK_ENVIRONMENT=YOUR_PUBLISHING_ENVIRONMENT
 
 # For live preview 
-CONTENTSTACK_MANAGEMENT_TOKEN= 
-CONTENTSTACK_API_HOST=api.contentstack.io
+CONTENTSTACK_PREVIEW_TOKEN=
 CONTENTSTACK_APP_HOST=app.contentstack.com
+CONTENTSTACK_PREVIEW_HOST=rest-preview.contentstack.com
 CONTENTSTACK_LIVE_PREVIEW=true
 CONTENTSTACK_LIVE_EDIT_TAGS=true
