+++ README.md
@@ -1,32 +1,47 @@
-[![Contentstack Logo](/public/contentstack-readme-logo.png)](https://www.contentstack.com/)
 
+# academy-upgrade
 
-# Build a Starter Website with Next.js and Contentstack
+This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
 
-About Contentstack: Contentstack is a headless CMS with an API-first approach that puts content at the centre. It is designed to simplify the process of publication by separating code from content.
+## Getting Started
 
-About this project: Next.js is a minimalistic framework for server-rendered React applications.This guide will help you create a starter website built on top of Next.js with minimal steps.
+First, run the development server:
 
+```bash
+npm run dev
+# or
+yarn dev
+# or
+pnpm dev
+# or
+bun dev
+```
 
+Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
 
-![Contentstack-Nextjs-starter-app](/public/starter-app.png)
+You can start editing the page by modifying `app/page.tsx`. The page auto-updates as you edit the file.
 
-## Live Demo
+This project uses [`next/font`](https://nextjs.org/docs/basic-features/font-optimization) to automatically optimize and load Inter, a custom Google Font.
 
-You can check the [live demo](https://contentstack-nextjs-starter-app.vercel.app) to get first-hand experience of the website.
+## Learn More
 
+To learn more about Next.js, take a look at the following resources:
 
-## Tutorial
+- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
+- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
 
-We have created an in-depth tutorial on how you can create a Next.js starter website using Contentstack's Node.js SDK and its fetch content from Contentstack.
+You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js/) - your feedback and contributions are welcome!
 
-[Build Website using Next.js and Contentstack](https://www.contentstack.com/docs/developers/sample-apps/build-a-starter-website-using-next-js-and-contentstack/)
+## Deploy on Vercel
 
+The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
 
-**More Resources**
+Check out our [Next.js deployment documentation](https://nextjs.org/docs/deployment) for more details.
 
-Read Contentstack [docs](https://www.contentstack.com/docs/)
-
-Region support [docs](https://www.contentstack.com/docs/developers/selecting-region-in-contentstack-starter-apps)
-
-Learn about [Next.js](https://learnnextjs.com/)
+>>>>>>> da15541 (Initial commit from Create Next App)
+>>>>>>>
+>>>>>>
+>>>>>
+>>>>
+>>>
+>>
