+++ app/globals.css
@@ -1,0 +1,15 @@
+html,
+body {
+  font-family: Inter, sans-serif;
+  font-size: 14px;
+  margin: 0;
+  line-height: 22px;
+}
+a {
+  color: inherit;
+  text-decoration: none;
+}
+
+* {
+  box-sizing: border-box;
+}
