+++ components/archive-relative.tsx
@@ -22,7 +22,7 @@ export default function ArchiveRelative({ blogs }: BlogListProps) {
   return (
     <>
       {blogs?.map((blog, idx) => (
-        <Link href={blog.url} key={idx}>
+        <Link legacyBehavior href={blog.url} key={idx}>
           <a>
             <div>
               <h4 {...blog.$?.title as {}}>{blog.title}</h4>
