+++ components/blog-section.tsx
@@ -48,7 +48,7 @@ export default function BlogSection(props: FeaturedBlogProps) {
           <h2 {...fromBlog.$?.title_h2 as {}}>{fromBlog.title_h2}</h2>
         )}
         {fromBlog.view_articles && (
-          <Link href={fromBlog.view_articles.href}>
+          <Link legacyBehavior href={fromBlog.view_articles.href}>
             <a
               className='btn secondary-btn article-btn'
               {...fromBlog.view_articles.$?.title}
@@ -75,7 +75,7 @@ export default function BlogSection(props: FeaturedBlogProps) {
                 <div>{parse(blog.body.slice(0, 300))}</div>
               )}
               {blog.url && (
-                <Link href={blog.url} passHref>
+                <Link legacyBehavior href={blog.url} passHref>
                   <a className='blogpost-readmore'>{'Read More -->'}</a>
                 </Link>
               )}
