+++ components/card-section.tsx
@@ -27,7 +27,7 @@ export default function CardSection({ cards }: CardProps) {
           {card.description && <p {...card.$?.description as {}}>{card.description}</p>}
           <div className='card-cta'>
             {card.call_to_action.title && card.call_to_action.href && (
-              <Link href={card.call_to_action.href}>
+              <Link legacyBehavior href={card.call_to_action.href}>
                 <a className='btn primary-btn'>{card.call_to_action.title}</a>
               </Link>
             )}
