+++ components/hero-banner.tsx
@@ -57,7 +57,7 @@ export default function HeroBanner(props: BannerProps) {
           ''
         )}
         {banner.call_to_action.title && banner.call_to_action.href ? (
-          <Link href={banner?.call_to_action.href}>
+          <Link legacyBehavior href={banner?.call_to_action.href}>
             <a className='btn tertiary-btn' {...banner.call_to_action.$?.title}>
               {banner?.call_to_action.title}
             </a>
