+++ contentstack-sdk/index.js
@@ -2,36 +2,33 @@ import * as contentstack from 'contentstack';
 import * as Utils from '@contentstack/utils';
 
 import ContentstackLivePreview from '@contentstack/live-preview-utils';
-import getConfig from 'next/config';
-
-const { publicRuntimeConfig } = getConfig();
-const envConfig = process.env.CONTENTSTACK_API_KEY
-  ? process.env
-  : publicRuntimeConfig;
 
 const Stack = contentstack.Stack({
-  api_key: envConfig.CONTENTSTACK_API_KEY
-    ? envConfig.CONTENTSTACK_API_KEY
-    : envConfig.NEXT_PUBLIC_CONTENTSTACK_API_KEY,
-  delivery_token: envConfig.CONTENTSTACK_DELIVERY_TOKEN,
-  environment: envConfig.CONTENTSTACK_ENVIRONMENT,
-  region: envConfig.CONTENTSTACK_REGION ? envConfig.CONTENTSTACK_REGION : 'us',
+  api_key: process.env.CONTENTSTACK_API_KEY
+    ? process.env.CONTENTSTACK_API_KEY
+    : process.env.NEXT_PUBLIC_CONTENTSTACK_API_KEY,
+  delivery_token: process.env.CONTENTSTACK_DELIVERY_TOKEN,
+  environment: process.env.CONTENTSTACK_ENVIRONMENT,
+  region: process.env.CONTENTSTACK_REGION ? process.env.CONTENTSTACK_REGION : 'us',
   live_preview: {
     enable: true,
-    management_token: envConfig.CONTENTSTACK_MANAGEMENT_TOKEN,
-    host: envConfig.CONTENTSTACK_API_HOST,
+    host: process.env.CONTENTSTACK_PREVIEW_HOST,
+    preview_token: process.env.CONTENTSTACK_PREVIEW_TOKEN
   },
 });
 
-if (envConfig.CONTENTSTACK_API_HOST) {
-  Stack.setHost(envConfig.CONTENTSTACK_API_HOST);
-}
-
 ContentstackLivePreview.init({
   stackSdk: Stack,
+  stackDetails: {
+    apiKey: process.env.CONTENTSTACK_API_KEY,
+    environment: process.env.CONTENTSTACK_ENVIRONMENT,
+    branch: process.env.CONTENTSTACK_BRANCH,
+},
+   
   clientUrlParams: {
-    host: envConfig.CONTENTSTACK_APP_HOST,
+    host: process.env.CONTENTSTACK_APP_HOST,
   },
+  enable: true,
   ssr: false,
 });
