+++ package.json
@@ -1,9 +1,7 @@
 {
   "name": "contentstack-nextjs-starter-app",
-  "description": "A starter app for Contentstack and Nextjs",
-  "version": "2.0.3",
+  "version": "0.1.0",
   "private": true,
-  "author": "Contentstack",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
@@ -11,27 +9,21 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@contentstack/live-preview-utils": "^1.2.1",
-    "@contentstack/utils": "^1.1.2",
-    "axios": "^1.3.5",
-    "contentstack": "^3.17.0",
-    "html-react-parser": "^3.0.4",
-    "moment": "^2.29.4",
-    "next": "^12.3.1",
-    "next-pwa": "^5.6.0",
-    "nprogress": "^0.2.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-json-view": "^1.21.3",
-    "react-loading-skeleton": "^3.1.0"
+    "@contentstack/live-preview-utils": "^2.0.3",
+    "contentstack": "^3.20.1",
+    "html-react-parser": "^5.1.12",
+    "moment": "^2.30.1",
+    "next": "14.0.4",
+    "react": "^18",
+    "react-dom": "^18",
+    "react-loading-skeleton": "^3.4.0"
   },
   "devDependencies": {
-    "@types/node": "^18.7.18",
-    "@types/nprogress": "^0.2.0",
-    "@types/react": "^18.0.20",
-    "@types/react-dom": "^18.0.6",
-    "eslint": "^8.23.1",
-    "eslint-config-next": "^12.3.1",
-    "typescript": "^4.8.3"
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "eslint": "^8",
+    "eslint-config-next": "14.2.5",
+    "typescript": "^5"
   }
 }
